Add a base form class for easier validation
